Factor out common NumericalChannel parts
